+++ tests/conftest.py
@@ -46,3 +46,4 @@ def mock_sumstat():
         }
     )
     return mock_data
+