+++ CHANGELOG.md
@@ -5,7 +5,18 @@
 ### Added
 
 - merge the overlapped independent loci (optional).
-- 
+
+### Changed
+
+
+### Fixed
+
+## [0.0.2] - 2022-12-22
+
+### Added
+
+- identify the independent lead snps by distance only
+- expand the independent lead snps to independent loci by given range.
 
 ### Changed